push

+++ Mergesort/Myfloor_func.c
@@ -1,5 +1,5 @@
 #include"Myfloor_func.h"
-int myfloor(float x)
+float myfloor(float x)
 {
     int a=(int)x;
     return(x>=0 || a==x)?a :a-1;
