+++ Mergesort/Myfloor_func.h
@@ -1,5 +1,5 @@
 #pragma once
-int myfloor(float x);
+float myfloor(float x);
 
 void mergesort(int a[],int l,int r);
 void merge(int a[],int l,int m,int r);
